Split PageComposer into Livewire concern traits
The god component is now a thin orchestrator. Four traits under
Livewire/Concerns/ carry the previously inline concerns:

- HandlesImageUploads
    photo, newsletter_image, slider_image properties + the saved /
    deleted listener pair + the field whitelist.

- HandlesTemplates
    templateName, selectedTemplate properties + saveTemplate,
    selectTemplate, loadTemplate.

- InteractsWithLanguages
    languages, currentLanguage, availableLanguages, selectableLanguages
    properties + addLanguage, setLanguage, setRowsLanguage,
    hydrateLanguages, copyContent, languageAdded listener.

- ManagesRows
    rows, showMiniMap properties + addRow, updateRowSorting,
    rowSortableKey, sortedRows computed, sortItems, deleteRow and
    rowUpdated listeners, plus the ensureUnsavedRowsHaveUuid /
    calculateRowAvailableSpace helpers (now protected so the host
    can still call them from setPageContent).

Cross-trait coupling (e.g. ManagesRows reads currentLanguage,
InteractsWithLanguages writes rows) is documented via @property
annotations on each trait.

What stays on PageComposer: mount, render, rules, syncPageState,
columnWidth, saveContent / updateContent / persistPage,
setPageContent (DB -> state hydration), small leaf listeners
(componentAdded, dateSelected, categorySelected, tagsUpdated),
and the cache() accessor.

PageComposer.php: 676 -> 307 lines (66% reduction from the original
902 at the start of this work).
