+++ src/PageComposer/Livewire/Concerns/HandlesImageUploads.php
@@ -1,0 +1,43 @@
+<?php
+
+namespace Flobbos\PageComposer\Livewire\Concerns;
+
+use Livewire\Attributes\On;
+
+/**
+ * @property array|null $page
+ */
+trait HandlesImageUploads
+{
+    private const PHOTO_FIELDS = ['photo', 'newsletter_image', 'slider_image'];
+
+    public $photo;
+    public $newsletter_image;
+    public $slider_image;
+
+    #[On('eventImageUploadComponentSaved.pageComposer.mainPhoto')]
+    #[On('eventImageUploadComponentSaved.pageComposer.newsletterImage')]
+    #[On('eventImageUploadComponentSaved.pageComposer.sliderImage')]
+    public function imageSaved(string $field, string $imagePath, ?int $itemIndex = null): void
+    {
+        $this->setPhotoField($field, $imagePath);
+    }
+
+    #[On('eventImageUploadComponentDeleted.pageComposer.mainPhoto')]
+    #[On('eventImageUploadComponentDeleted.pageComposer.newsletterImage')]
+    #[On('eventImageUploadComponentDeleted.pageComposer.sliderImage')]
+    public function imageDeleted(string $field, ?string $imagePath = null, ?int $itemIndex = null): void
+    {
+        $this->setPhotoField($field, null);
+    }
+
+    private function setPhotoField(string $field, ?string $value): void
+    {
+        if (!in_array($field, self::PHOTO_FIELDS, true)) {
+            return;
+        }
+
+        $this->{$field} = $value;
+        $this->page[$field] = $value;
+    }
+}
