+++ src/PageComposer/Livewire/PageComposer.php
@@ -5,53 +5,40 @@ namespace Flobbos\PageComposer\Livewire;
 use Exception;
 use Livewire\Component;
 use Illuminate\Support\Arr;
-use Illuminate\Support\Str;
-use Flobbos\PageComposer\Models\Row;
+use Flobbos\PageComposer\Livewire\Concerns\HandlesImageUploads;
+use Flobbos\PageComposer\Livewire\Concerns\HandlesTemplates;
+use Flobbos\PageComposer\Livewire\Concerns\InteractsWithLanguages;
+use Flobbos\PageComposer\Livewire\Concerns\ManagesRows;
 use Flobbos\PageComposer\Models\Page;
-use Flobbos\PageComposer\Models\Language;
 use Flobbos\PageComposer\Models\PageTemplate;
 use Flobbos\PageComposer\Services\PageBuilder;
 use Flobbos\PageComposer\Services\PageBuilderResult;
 use Flobbos\PageComposer\Services\PageComposerCache;
-use Flobbos\PageComposer\Services\SortService;
 use Livewire\Attributes\On;
-use Livewire\Attributes\Computed;
 
 class PageComposer extends Component
 {
-    //page
+    use HandlesImageUploads;
+    use HandlesTemplates;
+    use InteractsWithLanguages;
+    use ManagesRows;
+
     public $elements, $page, $pageId, $pageCategory;
     public $pageTags = [];
-
-    //language
-    public $languages, $currentLanguage, $availableLanguages, $selectableLanguages;
-
-    //photos
-    public $photo, $newsletter_image, $slider_image;
-
-    //options
     public $categories, $tags;
 
-    //error handling
     public $exceptionMessage;
     public $showErrorMessage = false;
 
-    //Settings
     public $settingsBox;
-    public $showMiniMap = false;
     public $currentElement = ['name' => ''];
 
-    //Preview
     public $previewMode = true;
     public $previewLanguage;
 
-    public $rows = [];
     public $pageTranslations = [];
 
     public $displayDate, $publishedOn;
-
-    //Template name
-    public $templateName, $selectedTemplate;
 
     protected function rules()
     {
@@ -112,134 +99,6 @@ class PageComposer extends Component
         ]);
 
         return Arr::get($sizes, $size, 'w-full');
-    }
-
-    /**
-     * Add a new language to the content
-     *
-     * @param integer $language_id
-     * @return void
-     */
-    public function addLanguage(int $language_id): void
-    {
-        $selectedLanguage = $this->cache()->languages()->firstWhere('id', $language_id) ?? Language::find($language_id);
-
-        if (!$selectedLanguage) {
-            return;
-        }
-
-        //Set current active language if not present
-        if (empty($this->currentLanguage)) {
-            $this->currentLanguage = $selectedLanguage;
-        }
-
-        //if (!$this->pageId) {
-        $this->pageTranslations[$selectedLanguage->locale]['language_id'] = $selectedLanguage->id;
-        $this->setRowsLanguage($selectedLanguage->id);
-        //}
-
-        $this->hydrateLanguages();
-    }
-
-    /**
-     * Set the current language for the content
-     *
-     * @param integer $language_id
-     * @return void
-     */
-    public function setLanguage(int $language_id): void
-    {
-        $this->currentLanguage = $this->cache()->languages()->firstWhere('id', $language_id) ?? Language::find($language_id);
-    }
-
-    /**
-     * Add a new row of content
-     *
-     * @return void
-     */
-    public function addRow(): void
-    {
-        $this->rows[$this->currentLanguage->locale]['rows'][] = [
-            'uuid' => (string) Str::uuid(),
-            'columns' => [],
-            'attributes' => [],
-            'alignment' => 'center',
-            'expanded' => false,
-            'active' => true,
-            'sorting' => $this->rows[$this->currentLanguage->locale]['rows'] ? count($this->rows[$this->currentLanguage->locale]['rows']) + 1 : 1,
-            'available_space' => 12
-        ];
-    }
-
-    /**
-     * Update the sorting for the rows. Called by Livewire 4's wire:sort when
-     * a row is dropped in a new position in the mini map.
-     *
-     * @param string|int $id       sortable key from wire:sort:item (uuid, db id, or tmp-N)
-     * @param int        $position zero-based target position
-     */
-    public function updateRowSorting($id, $position): void
-    {
-        $locale = $this->currentLanguage->locale ?? null;
-        if (!$locale) {
-            return;
-        }
-
-        $rows = $this->rows[$locale]['rows'] ?? [];
-
-        $this->rows[$locale]['rows'] = app(SortService::class)->reorder(
-            $rows,
-            fn(array $row, $key) => $this->rowSortableKey($row, (int) $key),
-            $id,
-            (int) $position,
-        );
-    }
-
-    public function rowSortableKey(array $row, int $fallbackIndex): string
-    {
-        if (filled(Arr::get($row, 'id'))) {
-            return (string) Arr::get($row, 'id');
-        }
-
-        if (filled(Arr::get($row, 'uuid'))) {
-            return (string) Arr::get($row, 'uuid');
-        }
-
-        return 'tmp-' . $fallbackIndex;
-    }
-
-    /**
-     * Computed property to get sorted rows
-     *
-     * @return array
-     */
-    #[Computed]
-    public function sortedRows(): array
-    {
-        if (!isset($this->currentLanguage)) {
-            return [];
-        }
-
-        if (empty($this->rows[$this->currentLanguage->locale]['rows'])) {
-            return [];
-        }
-
-        return Arr::sort($this->rows[$this->currentLanguage->locale]['rows'], function ($value) {
-            return $value['sorting'];
-        });
-    }
-
-    /**
-     * Sort column items
-     *
-     * @param array $items
-     * @return array
-     */
-    public function sortItems(array $items): array
-    {
-        return Arr::sort($items, function ($value) {
-            return $value['sorting'];
-        });
     }
 
     /**
@@ -312,109 +171,7 @@ class PageComposer extends Component
         );
     }
 
-    /**
-     * Copy all content from one language to another
-     *
-     * @param string $source source language
-     * @param string $target target language
-     * @return void
-     */
-    public function copyContent(string $source, string $target): void
-    {
-        $this->rows[$target] = $this->rows[$source];
-        //Remove database IDs from copied content
-        foreach ($this->rows[$target]['rows'] as $rowKey => $row) {
-            Arr::forget($this->rows, $target . '.rows.' . $rowKey . '.id');
-            $this->rows[$target]['rows'][$rowKey]['uuid'] = (string) Str::uuid();
-            foreach (Arr::get($row, 'columns', []) as $columnKey => $column) {
-                Arr::forget($this->rows, $target . '.rows.' . $rowKey . '.columns.' . $columnKey . '.id');
-                foreach (Arr::get($column, 'column_items', []) as $itemKey => $item) {
-                    Arr::forget($this->rows, $target . '.rows.' . $rowKey . '.columns.' . $columnKey . '.column_items.' . $itemKey . '.id');
-                }
-            }
-        }
-        $language = Language::where('locale', $target)->first();
-        $this->addLanguage($language->id);
-    }
-
-    /**
-     * Save current page as template
-     *
-     * @return void
-     */
-    public function saveTemplate(): void
-    {
-        $this->validate([
-            'templateName' => 'required'
-        ], ['required' => '(required)']);
-
-        //Set languages
-        $languages = [];
-
-        foreach ($this->pageTranslations as $trans) {
-            $languages[] = Arr::get($trans, 'language_id');
-        }
-
-        //Set content structure
-        $content = $this->rows;
-        foreach ($content as $langKey => $lang) {
-            foreach (Arr::get($lang, 'rows', []) as $key => $row) {
-                foreach (Arr::get($row, 'columns', []) as $columnKey => $column) {
-                    foreach (Arr::get($column, 'column_items', []) as $itemKey => $item) {
-                        $content[$langKey]['rows'][$key]['columns'][$columnKey]['column_items'][$itemKey]['content'] = [];
-                    }
-                }
-            }
-        }
-        //Create template
-        PageTemplate::create([
-            'name' => $this->templateName,
-            'content' => $content,
-            'languages' => $languages,
-            'user_id' => auth()->id()
-        ]);
-        $this->resetErrorBag();
-        $this->reset('templateName');
-        session()->flash('template_saved', '(saved)');
-    }
-
-    public function selectTemplate()
-    {
-        return redirect()->route('page-composer::pages.create', ['template' => $this->selectedTemplate]);
-    }
-
-    public function loadTemplate($templateId)
-    {
-        //Load template informaiton
-        $template = PageTemplate::find($templateId);
-        //Load languages
-        $languages = Language::whereIn('id', $template->languages)->get();
-        //Set page translations
-        foreach ($languages as $lang) {
-            $this->pageTranslations[$lang->locale] = [];
-            $this->addLanguage($lang->id);
-        }
-
-        //Set rows and columns
-        foreach ($template->content as $key => $rows) {
-            $this->rows[$key] = $rows;
-            $this->ensureUnsavedRowsHaveUuid($key);
-        }
-    }
-
     /******* Listeners *******/
-    /**
-     * Listener for when a new language has been added
-     *
-     * @return void
-     */
-    #[On('languageAdded')]
-    public function languageAdded(): void
-    {
-        // Language can be added during editing, so refresh this cache explicitly.
-        $this->cache()->languages(true);
-        $this->hydrateLanguages();
-    }
 
     /**
      * Refresh component list when a new one has been generated
@@ -451,37 +208,6 @@ class PageComposer extends Component
     {
         $this->pageCategory = $option;
         $this->page['category_id'] = Arr::get($option, 'id');
-    }
-
-    /**
-     * Event listener for when a row has been removed
-     *
-     * @param string $row_key
-     * @return void
-     */
-    #[On('deleteRow')]
-    public function deleteRow(string $rowKey): void
-    {
-        if (isset($this->rows[$this->currentLanguage->locale]['rows'][$rowKey]['id'])) {
-            if ($row = Row::find($this->rows[$this->currentLanguage->locale]['rows'][$rowKey]['id'])) {
-                $row->delete();
-            }
-        }
-        unset($this->rows[$this->currentLanguage->locale]['rows'][$rowKey]);
-    }
-
-    /**
-     * Event listener for an update to the row data
-     *
-     * @param array $row
-     * @param string $rowKey
-     * @return void
-     */
-    #[On('rowUpdated')]
-    #[On('columnUpdated')]
-    public function rowUpdated(array $row, string $rowKey): void
-    {
-        $this->rows[$this->currentLanguage->locale]['rows'][$rowKey] = $row;
     }
 
     /**
@@ -538,6 +264,7 @@ class PageComposer extends Component
                 $rowData['available_space'] = $this->calculateRowAvailableSpace(Arr::get($rowData, 'columns', []));
                 $this->rows[$row->language->locale]['rows'][] = $rowData;
             }
+
             //Set element data
             foreach ($this->rows as $lang => $langRow) {
                 foreach ($langRow['rows'] as $rowKey => $row) {
@@ -577,100 +304,4 @@ class PageComposer extends Component
         }
     }
 
-    /**
-     * Hydrate language specific arrays
-     *
-     * @return void
-     */
-    public function hydrateLanguages(): void
-    {
-        //All languages
-        $this->languages = $this->cache()->languages();
-        $this->availableLanguages = collect();
-        $this->selectableLanguages = $this->languages;
-        //Available languages in article
-
-        foreach ($this->pageTranslations as $trans) {
-            if (isset($trans['language_id'])) {
-                $this->availableLanguages->push($this->languages->where('id', $trans['language_id'])->first());
-            }
-        }
-
-        //Languages that can be added
-        foreach ($this->selectableLanguages as $key => $lang) {
-            if ($this->availableLanguages->contains('id', $lang->id)) {
-                $this->selectableLanguages->forget($key);
-            }
-        }
-    }
-
-    /**
-     * Set current language for the rows
-     *
-     * @param integer $language_id
-     * @return void
-     */
-    public function setRowsLanguage(int $language_id)
-    {
-        $lang = $this->cache()->languages()->firstWhere('id', $language_id) ?? Language::find($language_id);
-
-        if (!$lang) {
-            return;
-        }
-
-        if (! isset($this->rows[$lang->locale])) {
-            $this->rows[$lang->locale] = [
-                'rows' => [],
-            ];
-        }
-
-        $this->ensureUnsavedRowsHaveUuid($lang->locale);
-    }
-
-    private function ensureUnsavedRowsHaveUuid(string $locale): void
-    {
-        foreach (Arr::get($this->rows, $locale . '.rows', []) as $rowKey => $row) {
-            $this->rows[$locale]['rows'][$rowKey]['available_space'] = $this->calculateRowAvailableSpace(Arr::get($row, 'columns', []));
-
-            if (filled(Arr::get($row, 'id')) || filled(Arr::get($row, 'uuid'))) {
-                continue;
-            }
-
-            $this->rows[$locale]['rows'][$rowKey]['uuid'] = (string) Str::uuid();
-        }
-    }
-
-    private function calculateRowAvailableSpace(array $columns): int
-    {
-        return max(0, 12 - (int) collect($columns)
-            ->sum(fn($column) => (int) Arr::get($column, 'column_size', 0)));
-    }
-
-    private const PHOTO_FIELDS = ['photo', 'newsletter_image', 'slider_image'];
-
-    #[On('eventImageUploadComponentSaved.pageComposer.mainPhoto')]
-    #[On('eventImageUploadComponentSaved.pageComposer.newsletterImage')]
-    #[On('eventImageUploadComponentSaved.pageComposer.sliderImage')]
-    public function imageSaved(string $field, string $imagePath, ?int $itemIndex = null): void
-    {
-        $this->setPhotoField($field, $imagePath);
-    }
-
-    #[On('eventImageUploadComponentDeleted.pageComposer.mainPhoto')]
-    #[On('eventImageUploadComponentDeleted.pageComposer.newsletterImage')]
-    #[On('eventImageUploadComponentDeleted.pageComposer.sliderImage')]
-    public function imageDeleted(string $field, ?string $imagePath = null, ?int $itemIndex = null): void
-    {
-        $this->setPhotoField($field, null);
-    }
-
-    private function setPhotoField(string $field, ?string $value): void
-    {
-        if (!in_array($field, self::PHOTO_FIELDS, true)) {
-            return;
-        }
-
-        $this->{$field} = $value;
-        $this->page[$field] = $value;
-    }
 }
